requirements.txt
asgiref==3.5.0
bidict==0.22.0
Django==4.0.3
dnspython==2.2.1
eventlet==0.33.0
greenlet==1.1.2
python-engineio==4.3.1
python-socketio==5.5.2
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
